+++ CHANGELOG.md
@@ -110,6 +110,7 @@ and this project adheres to
 - cosmwasm-std: Rename type `KV` to `Pair` in order to comply to naming
   convention as enforced by clippy rule `upper_case_acronyms` from Rust 1.51.0
   on.
+- cosmwasm-std: `ContractInfo::address` is now of type `Addr`.
 - cosmwasm-vm: Bump required marker export `cosmwasm_vm_version_4` to
   `interface_version_5`.
 - cosmwasm-vm: Rename trait `Api` to `BackendApi` to better express this is the